requirements.txt
fastapi==0.115.0
     uvicorn==0.30.6
     pdfminer.six==20231228
     python-multipart==0.0.9